move files to handlers and models folders
